core: Allow -var flags to contain lists and maps
This commit applies the same logic as the previous one in order to allow
-var flags to contain list and map literals written in HCL. In addition,
it merges map keys, and moves up the logic applying variable defaults to
the top level such that all overrides are in one location in code.

The override syntax has changed (to be noted in CHANGELOG as a breaking
change), so several tests needed their syntax updating from the old
`amis.us-east-1 = "newValue"` style to `amis = "{ "us-east-1" =
"newValue"}"` style as defined in TF-002.